Strip out `SpacefinderOkr2ImagesLoaded` logic
